Experiment with BuildableReferences
