--- api/src/main/java/io/strimzi/api/kafka/model/ReplicatedJvmPods.java
@@ -1,159 +1,0 @@
-/*
- * Copyright 2018, Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.api.kafka.model;
-
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.fabric8.kubernetes.api.model.Affinity;
-import io.strimzi.crdgenerator.annotations.Description;
-import io.strimzi.crdgenerator.annotations.KubeLink;
-import io.strimzi.crdgenerator.annotations.Minimum;
-
-import io.fabric8.kubernetes.api.model.Toleration;
-import io.sundr.builder.annotations.Buildable;
-
-import java.io.Serializable;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * A baseclass for {@link Kafka} and {@link Zookeeper} with all their common properties.
- */
-@Buildable(editableEnabled = false,
-        generateBuilderPackage = true,
-        builderPackage = "io.strimzi.api.kafka.model"
-)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public abstract class ReplicatedJvmPods implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private int replicas;
-
-    private String image;
-
-    private Resources resources;
-
-    private Probe livenessProbe;
-
-    private Probe readinessProbe;
-
-    private JvmOptions jvmOptions;
-
-    private Map<String, Object> metrics = new HashMap<>(0);
-
-    private Affinity affinity;
-
-    private List<Toleration> tolerations;
-
-    private Map<String, Object> additionalProperties = new HashMap<>(0);
-
-    @Description("The number of pods in the cluster.")
-    @Minimum(1)
-    @JsonProperty(required = true)
-    public int getReplicas() {
-        return replicas;
-    }
-
-    public void setReplicas(int replicas) {
-        this.replicas = replicas;
-    }
-
-    @Description("The docker image for the pods.")
-    @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    @Description("Resource constraints (limits and requests).")
-    public Resources getResources() {
-        return resources;
-    }
-
-    public void setResources(Resources resources) {
-        this.resources = resources;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    @Description("Pod liveness checking.")
-    public Probe getLivenessProbe() {
-        return livenessProbe;
-    }
-
-    public void setLivenessProbe(Probe livenessProbe) {
-        this.livenessProbe = livenessProbe;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-    @Description("Pod readiness checking.")
-    public Probe getReadinessProbe() {
-        return readinessProbe;
-    }
-
-    public void setReadinessProbe(Probe readinessProbe) {
-        this.readinessProbe = readinessProbe;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    @Description("JVM Options for pods")
-    public JvmOptions getJvmOptions() {
-        return jvmOptions;
-    }
-
-    public void setJvmOptions(JvmOptions jvmOptions) {
-        this.jvmOptions = jvmOptions;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    @Description("The Prometheus JMX Exporter configuration. " +
-            "See https://github.com/prometheus/jmx_exporter for details of the structure of this configuration.")
-    public Map<String, Object> getMetrics() {
-        return metrics;
-    }
-
-    public void setMetrics(Map<String, Object> metrics) {
-        this.metrics = metrics;
-    }
-
-    @Description("Pod affinity rules.")
-    @KubeLink(group = "core", version = "v1", kind = "affinity")
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    public Affinity getAffinity() {
-        return affinity;
-    }
-
-    public void setAffinity(Affinity affinity) {
-        this.affinity = affinity;
-    }
-
-    @Description("Pod's tolerations.")
-    @KubeLink(group = "core", version = "v1", kind = "tolerations")
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    public List<Toleration> getTolerations() {
-        return tolerations;
-    }
-
-    public void setTolerations(List<Toleration> tolerations) {
-        this.tolerations = tolerations;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-}
